Fix shame in private chats
If !shame is used in a private chat, that private chat does not
appear in the Map storing the chanells. The !shame would then
be able to kill the bot since we assumed the chat where the
command is issued to always exist.

Fixed the wrong assumption by using fromMaybe instead of fromJust.