Refactor action handling - handle state actions
Now handles messages in a way that all reactions to them are
prepared first and then run one after another. This allows
handling state-changing and state-keeping message reactions
in one function, independedly of the type of the original
message (whether it was a channel/server/private/ping message).

FIXME: this breaks the check whether the botname was mentioned